@@ -1,40 +1,45 @@
 source 'https://rubygems.org'
-ruby '2.0.0'
+
+# Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
 gem 'rails', '4.0.0'
+
+# Use sqlite3 as the database for Active Record
 gem 'sqlite3'
+
+# Use SCSS for stylesheets
 gem 'sass-rails', '~> 4.0.0'
+
+# Use Uglifier as compressor for JavaScript assets
 gem 'uglifier', '>= 1.3.0'
+
+# Use CoffeeScript for .js.coffee assets and views
 gem 'coffee-rails', '~> 4.0.0'
+
+# See https://github.com/sstephenson/execjs#readme for more supported runtimes
+# gem 'therubyracer', platforms: :ruby
+
+# Use jquery as the JavaScript library
 gem 'jquery-rails'
+
+# Turbolinks makes following links in your web application faster. Read more: https://github.com/rails/turbolinks
 gem 'turbolinks'
+
+# Build JSON APIs with ease. Read more: https://github.com/rails/jbuilder
 gem 'jbuilder', '~> 1.2'
-gem 'bootstrap-sass'
-gem 'cancan'
-gem 'devise'
-gem 'devise_invitable'
-gem 'figaro'
-gem 'haml-rails'
-gem 'rolify'
-gem 'simple_form', '>= 3.0.0.rc'
-group :assets do
-  gem 'therubyracer', :platform=>:ruby
+
+group :doc do
+  # bundle exec rake doc:rails generates the API under doc/api.
+  gem 'sdoc', require: false
 end
-group :development do
-  gem 'better_errors'
-  gem 'binding_of_caller', :platforms=>[:mri_19, :mri_20, :rbx]
-  gem 'html2haml'
-  gem 'hub', :require=>nil
-  gem 'quiet_assets'
-end
-group :development, :test do
-  gem 'rspec-rails'
-  gem 'thin'
-end
-group :production do
-  gem 'unicorn'
-end
-group :test do
-  gem 'capybara'
-  gem 'database_cleaner', '1.0.1'
-  gem 'email_spec'
-end
+
+# Use ActiveModel has_secure_password
+# gem 'bcrypt-ruby', '~> 3.0.0'
+
+# Use unicorn as the app server
+# gem 'unicorn'
+
+# Use Capistrano for deployment
+# gem 'capistrano', group: :development
+
+# Use debugger
+# gem 'debugger', group: [:development, :test]
